requirements.txt
langchain
python-dotenv
ipykernel
pypdf
beautifulsoup4
arxiv
pymupdf
wikipedia
langchain-openai 
sentence-transformers
langchain-huggingface
faiss-cpu
langchain-Chroma
langchain-ollama
streamlit
langchain-groq
fastapi
uvicorn
langserve
langchain-community
SQLAlchemy
mysql-connector-python
validators
youtube_transcript_api
pytube
unstructured
numexpr
huggingface-hub
gradio
duckduckgo_search